+++ parachain/pallets/ethereum-beacon-client/src/mock.rs
@@ -6,9 +6,7 @@ use crate as ethereum_beacon_client;
 use frame_support::parameter_types;
 use frame_system as system;
 use pallet_timestamp;
-use snowbridge_beacon_primitives::{
-	self as primitives, BeaconHeader, Fork, ForkVersions, PublicKey, Signature,
-};
+use primitives::{BeaconHeader, Fork, ForkVersions, PublicKey, Signature};
 use sp_core::H256;
 use sp_runtime::{
 	testing::Header,
@@ -206,8 +204,11 @@ pub fn new_tester<T: crate::Config>() -> sp_io::TestExternalities {
 	system::GenesisConfig::default().build_storage::<T>().unwrap().into()
 }
 
-pub struct BLSSignatureVerifyTest<const SYNC_COMMITTEE_SIZE: usize> {
-	pub sync_committee_bits: [u8; SYNC_COMMITTEE_SIZE],
+pub struct BLSSignatureVerifyTest<
+	const SYNC_COMMITTEE_SIZE: usize,
+	const SYNC_COMMITTEE_BITS_SIZE: usize,
+> {
+	pub sync_committee_bits: [u8; SYNC_COMMITTEE_BITS_SIZE],
 	pub sync_committee_signature: Signature,
 	pub pubkeys: [PublicKey; SYNC_COMMITTEE_SIZE],
 	pub header: BeaconHeader,
@@ -221,14 +222,17 @@ fn fixture_path(name: &str) -> PathBuf {
 
 fn initial_sync_from_file<const SYNC_COMMITTEE_SIZE: usize>(
 	name: &str,
-) -> primitives::InitialSync<SYNC_COMMITTEE_SIZE> {
+) -> primitives::InitialUpdate<SYNC_COMMITTEE_SIZE> {
 	let filepath = fixture_path(name);
 	serde_json::from_reader(File::open(&filepath).unwrap()).unwrap()
 }
 
-fn sync_committee_update_from_file<const SYNC_COMMITTEE_SIZE: usize>(
+fn sync_committee_update_from_file<
+	const SYNC_COMMITTEE_SIZE: usize,
+	const SYNC_COMMITTEE_BITS_SIZE: usize,
+>(
 	name: &str,
-) -> primitives::SyncCommitteeUpdate<SYNC_COMMITTEE_SIZE> {
+) -> primitives::SyncCommitteeUpdate<SYNC_COMMITTEE_SIZE, SYNC_COMMITTEE_BITS_SIZE> {
 	let filepath = fixture_path(name);
 	serde_json::from_reader(File::open(&filepath).unwrap()).unwrap()
 }
@@ -240,16 +244,22 @@ fn sync_committee_from_file<const SYNC_COMMITTEE_SIZE: usize>(
 	serde_json::from_reader(File::open(&filepath).unwrap()).unwrap()
 }
 
-fn finalized_header_update_from_file<const SYNC_COMMITTEE_SIZE: usize>(
+fn finalized_header_update_from_file<
+	const SYNC_COMMITTEE_SIZE: usize,
+	const SYNC_COMMITTEE_BITS_SIZE: usize,
+>(
 	name: &str,
-) -> primitives::FinalizedHeaderUpdate<SYNC_COMMITTEE_SIZE> {
+) -> primitives::FinalizedHeaderUpdate<SYNC_COMMITTEE_SIZE, SYNC_COMMITTEE_BITS_SIZE> {
 	let filepath = fixture_path(name);
 	serde_json::from_reader(File::open(&filepath).unwrap()).unwrap()
 }
 
-fn header_update_from_file<const SYNC_COMMITTEE_SIZE: usize>(
+fn header_update_from_file<
+	const SYNC_COMMITTEE_SIZE: usize,
+	const SYNC_COMMITTEE_BITS_SIZE: usize,
+>(
 	name: &str,
-) -> primitives::HeaderUpdate<SYNC_COMMITTEE_SIZE> {
+) -> primitives::HeaderUpdate<SYNC_COMMITTEE_SIZE, SYNC_COMMITTEE_BITS_SIZE> {
 	let filepath = fixture_path(name);
 	serde_json::from_reader(File::open(&filepath).unwrap()).unwrap()
 }
@@ -271,15 +281,17 @@ fn add_file_prefix(name: &str) -> String {
 }
 
 pub fn get_initial_sync<const SYNC_COMMITTEE_SIZE: usize>(
-) -> primitives::InitialSync<SYNC_COMMITTEE_SIZE> {
+) -> primitives::InitialUpdate<SYNC_COMMITTEE_SIZE> {
 	initial_sync_from_file::<SYNC_COMMITTEE_SIZE>(&add_file_prefix("initial_sync.json"))
 }
 
-pub fn get_committee_sync_period_update<const SYNC_COMMITTEE_SIZE: usize>(
-) -> primitives::SyncCommitteeUpdate<SYNC_COMMITTEE_SIZE> {
-	sync_committee_update_from_file::<SYNC_COMMITTEE_SIZE>(&add_file_prefix(
-		"sync_committee_update.json",
-	))
+pub fn get_committee_sync_period_update<
+	const SYNC_COMMITTEE_SIZE: usize,
+	const SYNC_COMMITTEE_BITS_SIZE: usize,
+>() -> primitives::SyncCommitteeUpdate<SYNC_COMMITTEE_SIZE, SYNC_COMMITTEE_BITS_SIZE> {
+	sync_committee_update_from_file::<SYNC_COMMITTEE_SIZE, SYNC_COMMITTEE_BITS_SIZE>(
+		&add_file_prefix("sync_committee_update.json"),
+	)
 }
 
 pub fn get_committee_sync_ssz_test_data<const SYNC_COMMITTEE_SIZE: usize>(
@@ -290,25 +302,34 @@ pub fn get_committee_sync_ssz_test_data<const SYNC_COMMITTEE_SIZE: usize>(
 	sync_committee_from_file::<SYNC_COMMITTEE_SIZE>(filename.as_str())
 }
 
-pub fn get_header_update<const SYNC_COMMITTEE_SIZE: usize>(
-) -> primitives::HeaderUpdate<SYNC_COMMITTEE_SIZE> {
-	header_update_from_file::<SYNC_COMMITTEE_SIZE>(&add_file_prefix("header_update.json"))
+pub fn get_header_update<
+	const SYNC_COMMITTEE_SIZE: usize,
+	const SYNC_COMMITTEE_BITS_SIZE: usize,
+>() -> primitives::HeaderUpdate<SYNC_COMMITTEE_SIZE, SYNC_COMMITTEE_BITS_SIZE> {
+	header_update_from_file::<SYNC_COMMITTEE_SIZE, SYNC_COMMITTEE_BITS_SIZE>(&add_file_prefix(
+		"header_update.json",
+	))
 }
 
-pub fn get_finalized_header_update<const SYNC_COMMITTEE_SIZE: usize>(
-) -> primitives::FinalizedHeaderUpdate<SYNC_COMMITTEE_SIZE> {
-	finalized_header_update_from_file::<SYNC_COMMITTEE_SIZE>(&add_file_prefix(
-		"finalized_header_update.json",
-	))
+pub fn get_finalized_header_update<
+	const SYNC_COMMITTEE_SIZE: usize,
+	const SYNC_COMMITTEE_BITS_SIZE: usize,
+>() -> primitives::FinalizedHeaderUpdate<SYNC_COMMITTEE_SIZE, SYNC_COMMITTEE_BITS_SIZE> {
+	finalized_header_update_from_file::<SYNC_COMMITTEE_SIZE, SYNC_COMMITTEE_BITS_SIZE>(
+		&add_file_prefix("finalized_header_update.json"),
+	)
 }
 
 pub fn get_validators_root<const SYNC_COMMITTEE_SIZE: usize>() -> H256 {
 	get_initial_sync::<SYNC_COMMITTEE_SIZE>().validators_root
 }
 
-pub fn get_bls_signature_verify_test_data<const SYNC_COMMITTEE_SIZE: usize>(
-) -> BLSSignatureVerifyTest<SYNC_COMMITTEE_SIZE> {
-	let finalized_update = get_finalized_header_update::<SYNC_COMMITTEE_SIZE>();
+pub fn get_bls_signature_verify_test_data<
+	const SYNC_COMMITTEE_SIZE: usize,
+	const SYNC_COMMITTEE_BITS_SIZE: usize,
+>() -> BLSSignatureVerifyTest<SYNC_COMMITTEE_SIZE, SYNC_COMMITTEE_BITS_SIZE> {
+	let finalized_update =
+		get_finalized_header_update::<SYNC_COMMITTEE_SIZE, SYNC_COMMITTEE_BITS_SIZE>();
 	let initial_sync = get_initial_sync::<SYNC_COMMITTEE_SIZE>();
 
 	BLSSignatureVerifyTest {
