+++ parachain/primitives/beacon/Cargo.toml
@@ -26,7 +26,6 @@ byte-slice-cast = { version = "1.2.1", default-features = false }
 
 snowbridge-ethereum = { path = "../../primitives/ethereum", default-features = false }
 static_assertions = { version = "1.1.0" }
-serde-hex = { version = "0.1.0", optional = true }
 
 [dev-dependencies]
 hex-literal = { version = "0.4.1" }
@@ -35,7 +34,6 @@ hex-literal = { version = "0.4.1" }
 default = ["std"]
 std = [
     "serde",
-    "serde-hex",
     "codec/std",
     "scale-info/std",
     "frame-support/std",
